+++ Xtellar Empire Navigation Operating System Control Unit Manager/src/com/generation/dao/IDAOPianeti.java
@@ -6,7 +6,7 @@ import com.generation.base.Entity;
 
 public interface IDAOPianeti {
 	
-	public int grandezzaMediaPianeti();
+	public double grandezzaMediaPianeti();
 	public int grandezzaTotalePianeti();
 	public String distanzaPianeti(String nomea, String nomeb);
 	public boolean doppione(String nome); //se un pianeta ha stesso nome di un altro
